// lib/gen/types/models/ConnectorSecurityRating.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { ModelsComputeBinding } from './ComputeBinding.ts'
import type { ModelsConnectionType } from './ConnectionType.ts'
import type { ModelsSecurityColor } from './SecurityColor.ts'
import type { ModelsSecurityLevel } from './SecurityLevel.ts'
import type { ModelsStartMode } from './StartMode.ts'

export type ModelsConnectorSecurityRating = {
  /**
   * @type string | undefined
   */
  accessUsageColor?: ModelsSecurityColor
  /**
   * @type string | undefined
   */
  auditColor?: ModelsSecurityColor
  /**
   * @type boolean | undefined
   */
  auditWorm?: boolean
  /**
   * @type boolean | undefined
   */
  autoPiiDeletion?: boolean
  /**
   * @type string | undefined
   */
  commColor?: ModelsSecurityColor
  /**
   * @type string | undefined
   */
  computeBinding?: ModelsComputeBinding
  /**
   * @description 追加型增强要素
   */
  connectionType?: ModelsConnectionType
  /**
   * @type string | undefined
   */
  connectorDid?: string
  /**
   * @type string | undefined
   */
  createdAt?: string
  /**
   * @type boolean | undefined
   */
  defaultPasswordChanged?: boolean
  /**
   * @type string | undefined
   */
  deletedAt?: string
  /**
   * @type boolean | undefined
   */
  didEnabled?: boolean
  /**
   * @type boolean | undefined
   */
  didSameChain?: boolean
  /**
   * @type string | undefined
   */
  gatingReason?: string
  /**
   * @type boolean | undefined
   */
  hsmTpm?: boolean
  /**
   * @type string | undefined
   */
  id?: string
  /**
   * @description 六维度颜色
   */
  identityColor?: ModelsSecurityColor
  /**
   * @type integer | undefined
   */
  logRetentionMonths?: number
  /**
   * @type boolean | undefined
   */
  measuredBoot?: boolean
  /**
   * @type boolean | undefined
   */
  mfaEnabled?: boolean
  /**
   * @type boolean | undefined
   */
  msgSigning?: boolean
  /**
   * @type string | undefined
   */
  overallLevel?: ModelsSecurityLevel
  /**
   * @description 汇总
   * @type number | undefined
   */
  overallScore?: number
  /**
   * @type string | undefined
   */
  platformColor?: ModelsSecurityColor
  /**
   * @type string | undefined
   */
  privacyColor?: ModelsSecurityColor
  /**
   * @type boolean | undefined
   */
  privacyMinimization?: boolean
  /**
   * @type boolean | undefined
   */
  secureBoot?: boolean
  /**
   * @type boolean | undefined
   */
  spvSupported?: boolean
  /**
   * @type string | undefined
   */
  startMode?: ModelsStartMode
  /**
   * @type string | undefined
   */
  thirdPartyCerts?: string
  /**
   * @type string | undefined
   */
  tlsVersion?: string
  /**
   * @type string | undefined
   */
  updatedAt?: string
  /**
   * @type boolean | undefined
   */
  usageControlEnforced?: boolean
}